+++ src/main/java/com/island/service/LifecycleService.java
@@ -5,7 +5,6 @@ import com.island.engine.SimulationWorld;
 import com.island.content.Animal;
 import com.island.content.Biomass;
 import com.island.content.DeathCause;
-import com.island.model.Cell;
 import com.island.util.RandomProvider;
 import java.util.concurrent.ExecutorService;